+++ ConfigManager.cs
@@ -1,39 +1,26 @@
 ﻿using BepInEx.Configuration;
-using System;
-using System.Collections.Generic;
-using System.Text;
 
 namespace ModListHashChecker
 {
     public class ConfigManager
     {
-        public static ConfigManager Instance { get; private set; }
-        public static ConfigFile config { get; private set; }
+        public static ConfigEntry<string> ExpectedModListHash { get; private set; } = null!;
+        public static ConfigEntry<bool> NoExpectedHashMessage { get; private set; } = null!;
+        public static ConfigEntry<bool> MenuWarning { get; private set; } = null!;
+        public static ConfigEntry<bool> JoinWarning { get; private set; } = null!;
+        public static ConfigEntry<string> WarningMessageText { get; private set; } = null!;
+        public static ConfigEntry<string> JoinWarningText { get; private set; } = null!;
+        public static ConfigEntry<int> JoinWarningDelay { get; private set; } = null!;
+        public static ConfigEntry<string> WarningButtonIgnoreText { get; private set; } = null!;
+        public static ConfigEntry<string> WarningButtonResetText { get; private set; } = null!;
+        public static ConfigEntry<string> NoHashMessageText { get; private set; } = null!;
+        public static ConfigEntry<string> NoHashRightButtonText { get; private set; } = null!;
+        public static ConfigEntry<string> NoHashLeftButtonText { get; private set; } = null!;
+        public static ConfigEntry<bool> DisplayHashOnLevelLoad { get; private set; } = null!;
 
-        public static void Init(ConfigFile config)
+        internal static void Init(ConfigFile config)
         {
-            Instance = new ConfigManager(config);
-        }
-
-        public static ConfigEntry<string> ExpectedModListHash { get; private set; }
-        public static ConfigEntry<bool> NoExpectedHashMessage { get; private set; }
-        //public static ConfigEntry<bool> AutosetModListHash { get; private set; }
-        public static ConfigEntry<bool> MenuWarning { get; private set; }
-        public static ConfigEntry<bool> JoinWarning { get; private set; }
-        public static ConfigEntry<string> WarningMessageText { get; private set; }
-        public static ConfigEntry<string> JoinWarningText { get; private set; }
-        public static ConfigEntry<int> JoinWarningDelay { get; private set; }
-        public static ConfigEntry<string> WarningButtonIgnoreText { get; private set; }
-        public static ConfigEntry<string> WarningButtonResetText { get; private set; }
-        public static ConfigEntry<string> NoHashMessageText { get; private set; }
-        public static ConfigEntry<string> NoHashRightButtonText { get; private set; }
-        public static ConfigEntry<string> NoHashLeftButtonText { get; private set; }
-
-        private ConfigManager(ConfigFile loadconfig)
-        {
-            config = loadconfig;
             ExpectedModListHash = config.Bind("General", "ExpectedModListHash", "", "The expected modlist hash for this modpack. Do not change this unless you know what you're doing.");
-            //AutosetModListHash = config.Bind("General", "AutosetModListHash", false, "If true, override the current expected hash to the current one. Do not change this unless you know what you're doing.");
             NoExpectedHashMessage = config.Bind("General", "NoExpectedHashMessage", true, "Enable or Disable displaying a warning message in the menus when the expected hash is empty. Do not change this unless you know what you're doing.");
             NoHashMessageText = config.Bind("Menu Warning", "NoHashMessageText", "ExpectedModListHash configuration item is blank.\n\nWould you like to set it to the currently loaded list of mods?", "Menu Message to display when the expected hash is empty");
             NoHashRightButtonText = config.Bind("Menu Warning", "NoHashRightButtonText", "No", "Button text for leaving the ExpectedModListHash blank");
@@ -45,6 +32,8 @@ namespace ModListHashChecker
             WarningMessageText = config.Bind("Menu Warning", "WarningMessageText", "Your modlist does not match the expected modlist hash.\n\n You may experience issues.", "Message to display in Hash Mismatch Menu Warning Message");
             WarningButtonIgnoreText = config.Bind("Menu Warning", "WarningButtonIgnoreText", "Okay", "Button text for ignoring the Hash Mismatch Menu Warning Message");
             WarningButtonResetText = config.Bind("Menu Warning", "WarningButtonResetText", "Reset", "Button text for reseting ExpectedModListHash to the detected hash in Hash Mismatch Menu Warning Message");
+
+            DisplayHashOnLevelLoad = config.Bind<bool>("General", "DisplayHashOnLevelLoad", true, "When enabled, will display the modlist hash in the chat on level load.");
         }
 
     }